requirements.txt
Flask
Flask-CORS
Flask-Migrate 
azure-storage-blob 
azure-core
gunicorn
langchain-community 
langchain-anthropic 
python-dotenv
mysql-connector-python
reportlab
Flask-SQLAlchemy
bcrypt
google-generativeai
pytesseract
langchain
logging
sqlalchemy